Added authentication checks on the endpoints and all tests completed apart from the auth/authz tests

+++ starter/test_app.py
@@ -7,6 +7,10 @@ from app import create_app
 
 from models import setup_db, Movie, Actor
 
+casting_assistant_token = 'Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IkVEaDJNMVFMcmV3WDdhMThsZFZzSyJ9.eyJpc3MiOiJodHRwczovL2Rldi1jNWs1YWs4NC5ldS5hdXRoMC5jb20vIiwic3ViIjoiYXV0aDB8NWY3OGY4MGE1ZGY0YmEwMDY5MjJlNmZkIiwiYXVkIjoiY2Fwc3RvbmUiLCJpYXQiOjE2MDE4MTg4ODcsImV4cCI6MTYwMTgyNjA4NywiYXpwIjoibVNHZEU0QUVGd3Rxd3VDeEoxTE10UGVaT1Z3ZzlyZEMiLCJzY29wZSI6IiIsInBlcm1pc3Npb25zIjpbInJlYWQ6YWN0b3JzIiwicmVhZDptb3ZpZXMiXX0.FEd3OP7eVu6HXG3KR0s3P6P-RJm_y573ZdtGaEKslwWGBXVeI3xg73h9V1lXrU7iwG7DxBxVbZGUow0WRbEHAIeth0n_n4YaPVIhQQ-2C6IeWD04vZnA4E89wrg-GVE1UB1Q8VFpDH7X0kzT6J-e8FrlwleWYay4PAQPS-8h9rxydPvLdn3KTWq5YBYPYlw5eyJEPHpcCNEd1q9w9CYAVN5v1SJmAgV9XLaVkPbD-YICxcugsuhJkfYa4KA4C7b0NBj2WBvYBLvvhii9zCj7hnQE_LaNlwBbyAH48PQHZeOpFuHqa3dTvdcQpplU_Ucbi4y9_Wz8f_Emc4pSvWBbaw'
+casting_director_token = 'Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IkVEaDJNMVFMcmV3WDdhMThsZFZzSyJ9.eyJpc3MiOiJodHRwczovL2Rldi1jNWs1YWs4NC5ldS5hdXRoMC5jb20vIiwic3ViIjoiYXV0aDB8NWY3OGY4M2ZhNmFmNjQwMDcxZDkzZGFlIiwiYXVkIjoiY2Fwc3RvbmUiLCJpYXQiOjE2MDE4MTg5NTcsImV4cCI6MTYwMTgyNjE1NywiYXpwIjoibVNHZEU0QUVGd3Rxd3VDeEoxTE10UGVaT1Z3ZzlyZEMiLCJzY29wZSI6IiIsInBlcm1pc3Npb25zIjpbImVkaXQ6YWN0b3JzIiwiZWRpdDptb3ZpZXMiLCJyZWFkOmFjdG9ycyIsInJlYWQ6bW92aWVzIl19.QG1X1ykMgPL6dtfIOdYKWHx1WteXqpuVK-CJy9eIb-RHVLd3WqE0AnreXgCHZjL0t7DyKZnqGW3i_HpaLtBhIdT064OBIaWj3F5oVMa52DfQf9I04qE9XGMjlb4gifn1lWr8nRZLBNlf19coWcVnGzKNe6C84VxvXbTdTrUrSnS3oo7GWxshSAsnBYR5xSjNJWuIvaAN37Cm1nXujeCr5iBbpoLUlFgsj5M3qqlI9vXhhV6s6feihuwFbaPdN9S7J8hK6Nh6PsHkfiv3QEASSLgoCV7dg47jHXIs14pY2hqQoaNqFA_C2WNMJhqIE8KztghPNCO93qYtKEAdTYUeIA'
+casting_executive_token = 'Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IkVEaDJNMVFMcmV3WDdhMThsZFZzSyJ9.eyJpc3MiOiJodHRwczovL2Rldi1jNWs1YWs4NC5ldS5hdXRoMC5jb20vIiwic3ViIjoiYXV0aDB8NWY3OGY4NTU1MmNiNTUwMDc4NDg1MzNiIiwiYXVkIjoiY2Fwc3RvbmUiLCJpYXQiOjE2MDE4MTkwNDcsImV4cCI6MTYwMTgyNjI0NywiYXpwIjoibVNHZEU0QUVGd3Rxd3VDeEoxTE10UGVaT1Z3ZzlyZEMiLCJzY29wZSI6IiIsInBlcm1pc3Npb25zIjpbImNyZWF0ZTphY3RvciIsImNyZWF0ZTptb3ZpZSIsImRlbGV0ZTphY3RvciIsImRlbGV0ZTptb3ZpZSIsImVkaXQ6YWN0b3JzIiwiZWRpdDptb3ZpZXMiLCJyZWFkOmFjdG9ycyIsInJlYWQ6bW92aWVzIl19.JN_tWPtAX6iZmKcx4pZo1WdVsxjKyiIHhoLKuFMNMjASYnuFFsWypucsAWon1Sjd_gXHZLDkZo3LAjHQ4uWgJPs--CipX2Syg5FnVwf5M2Bl6fom6hR6fZ9jSFCnym8VJQV1800r7NFTjZYegHtNJYWZ0KqvbywFadHoUpMlECS5HQwfFz-Sq8fLA13tmOYltvPHYWtmKoW442C8czSfXW1zh2is5eZkyFBOecVyHtENFpO9V-Pu7o9lf1KaLxRpwqomODls7JBq3Uvs6ruR5UBBEdqkNYI95hivFcDE_MoIj1-CSPW-q8kgq9zTPi5HFlppfsJLZ3zJRfZTUqqkOQ'
+
 class CapstoneProjectTests(unittest.TestCase):
 
     def setUp(self):
@@ -14,7 +18,7 @@ class CapstoneProjectTests(unittest.TestCase):
         self.app = create_app()
         self.client = self.app.test_client
         self.database_name = "capstone"
-        self.database_path = "postgres://{}/{}".format('postgres:@localhost:5432', self.database_name)
+        self.database_path = "postgres://{}/{}".format('postgres:Wtfpwnt12345!?@localhost:5432', self.database_name)
         setup_db(self.app, self.database_path)
 
         # binds the app to the current context
@@ -33,21 +37,21 @@ class CapstoneProjectTests(unittest.TestCase):
     # Get Tests
 
     def test_get_movies(self):
-        res = self.client().get('/movies')
+        res = self.client().get('/movies', headers = { 'Authorization': casting_assistant_token })
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
-        #self.assertTrue(data['movies'])
+        self.assertTrue(data['movies'])
 
 
     def test_get_actors(self):
-        res = self.client().get('/actors')
+        res = self.client().get('/actors', headers = { 'Authorization': casting_assistant_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
-        #self.assertTrue(data['actor'])
+        self.assertTrue(data['actors'])
 
     # Delete Tests
 
@@ -59,7 +63,7 @@ class CapstoneProjectTests(unittest.TestCase):
         actor = Actor.query.get(TestActor.id)
         actor_id = actor.id
 
-        res = self.client().delete(f'/actors/{actor_id}')
+        res = self.client().delete(f'/actors/{actor_id}', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -73,7 +77,7 @@ class CapstoneProjectTests(unittest.TestCase):
         movie = Movie.query.get(TestMovie.id)
         movie_id = movie.id
 
-        res = self.client().delete(f'/movies/{movie_id}')
+        res = self.client().delete(f'/movies/{movie_id}', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -82,14 +86,14 @@ class CapstoneProjectTests(unittest.TestCase):
 
     # Post Tests
     def test_create_actor(self):
-        res = self.client().post('/actors', json = {'name' : 'testing', 'surname':'testing', 'gender' : 'male', 'age':1})
+        res = self.client().post('/actors', json = {'name' : 'testing', 'surname':'testing', 'gender' : 'male', 'age':1}, headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_create_movie(self):
-        res = self.client().post('/movies', json = {'title' : 'testing', 'release_date':'01-01-01'})
+        res = self.client().post('/movies', json = {'title' : 'testing', 'release_date':'01-01-01'}, headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -98,14 +102,14 @@ class CapstoneProjectTests(unittest.TestCase):
     # Patch Tests
 
     def test_patch_movie(self):
-        res = self.client().patch('/movies/1/edit', json = {'title' : 'testier'})
+        res = self.client().patch('/movies/1/edit', json = {'title' : 'testier'}, headers = { 'Authorization': casting_director_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_patch_actor(self):
-        res = self.client().patch('/actors/1/edit', json = {'name' : 'testier'})
+        res = self.client().patch('/actors/1/edit', json = {'name' : 'testier'}, headers = { 'Authorization': casting_director_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -116,7 +120,7 @@ class CapstoneProjectTests(unittest.TestCase):
     # Delete non existent entry
 
     def test_404_delete_actor(self):
-        res = self.client().delete(f'/actors/5')
+        res = self.client().delete(f'/actors/5', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -124,7 +128,7 @@ class CapstoneProjectTests(unittest.TestCase):
         self.assertEqual(data['message'], 'not found')
 
     def test_404_delete_movie(self):
-        res = self.client().delete(f'/movies/5')
+        res = self.client().delete(f'/movies/5', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -133,7 +137,7 @@ class CapstoneProjectTests(unittest.TestCase):
 
     # update non-existent movie
     def test_404_update_movie(self):
-        res = self.client().patch(f'/movies/5/edit', json = {'title' : 'testier'})
+        res = self.client().patch(f'/movies/5/edit', json = {'title' : 'testier'} , headers = { 'Authorization': casting_director_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -143,7 +147,7 @@ class CapstoneProjectTests(unittest.TestCase):
     # update non-existent actor
 
     def test_404_update_actor(self):
-        res = self.client().patch(f'/actors/5/edit', json = {'surname' : 'testier'})
+        res = self.client().patch(f'/actors/5/edit', json = {'surname' : 'testier'}, headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -153,7 +157,7 @@ class CapstoneProjectTests(unittest.TestCase):
     # update movie without payload
 
     def test_422_update_movie(self):
-        res = self.client().patch(f'/movies/1/edit')
+        res = self.client().patch(f'/movies/1/edit', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -163,7 +167,7 @@ class CapstoneProjectTests(unittest.TestCase):
     # update actor without payload
 
     def test_422_update_actor(self):
-        res = self.client().patch(f'/actors/1/edit')
+        res = self.client().patch(f'/actors/1/edit', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -173,7 +177,7 @@ class CapstoneProjectTests(unittest.TestCase):
     # create actor without payload
 
     def test_422_create_actor(self):
-        res = self.client().post(f'/actors')
+        res = self.client().post(f'/actors', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -183,7 +187,7 @@ class CapstoneProjectTests(unittest.TestCase):
     # create actor without payload
 
     def test_422_create_movie(self):
-        res = self.client().post(f'/movies')
+        res = self.client().post(f'/movies', headers = { 'Authorization': casting_executive_token})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
